HCK-10586: enchanced ipv6 escaping logic
