+++ package.json
@@ -68,6 +68,7 @@
         "abort-controller": "3.0.0",
         "antlr4": "4.9.2",
         "async": "3.2.6",
+        "ip": "2.0.1",
         "lodash": "4.17.21",
         "node-fetch": "2.6.7"
     },
